A few more tweaks
